+++ pkg/sql/colmem/BUILD.bazel
@@ -22,9 +22,12 @@ go_library(
 go_test(
     name = "colmem_test",
     size = "small",
-    srcs = ["allocator_test.go"],
+    srcs = [
+        "allocator_test.go",
+        "reset_maybe_reallocate_test.go",
+    ],
+    embed = [":colmem"],
     deps = [
-        ":colmem",
         "//pkg/col/coldata",
         "//pkg/col/coldataext",
         "//pkg/settings/cluster",
